+++ src/main/java/theking530/staticpower/data/crafting/AbstractMachineRecipe.java
@@ -24,7 +24,7 @@ public abstract class AbstractMachineRecipe extends AbstractStaticPowerRecipe {
 	 * 
 	 * @return
 	 */
-	public long getPowerCost() {
+	public double getPowerCost() {
 		return processingSection.getPowerCost();
 	}
